style: move css files, fix classes, and imports
Moved CSS files to another folder.

Changed some of the class names to make it more uniform.

Added a few utility classes.

Moved some classes from internal to general.

Removed image modal styles in the garden of human rights css file.

Fixed sources in Navigation.js
