+++ src/app/(pages)/aboutEdBroadbent/page.js
@@ -2,7 +2,7 @@
 
 import "/src/app/globals.css";
 import "/src/app/css/pages/aboutEdBroadbent.css";
-import "/src/app/css/internal.css";
+import "/src/app/css/pages/internal.css";
 import "/src/app/css/imageModal.css"
 
 import Link from "next/link";
