+++ src/app/(pages)/amphitheatreAndStage/page.js
@@ -1,7 +1,7 @@
 "use client"
 
 import "/src/app/globals.css";
-import "/src/app/css/internal.css";
+import "/src/app/css/pages/internal.css";
 
 import Image from "next/image";
 import Link from "next/link";
@@ -69,11 +69,11 @@ export default function AmphitheatreAndStage() {
                 </section>
 
                 <section className="page-subsection u-content-width">
-                    <h2 className="page-subsection__title--events">Book an Event</h2>
+                    <h2 className="page-subsection__title mb-8">Book an Event</h2>
                     <p>Interested in hosting an event at Ed Broadbent Waterfront Park? Learn more about <span><a className="p-span-link" href="https://www.oshawa.ca/en/parks-recreation-and-culture/host-an-event.aspx" target="_blank" rel="noopener noreferrer">Hosting an Event</a></span> and <span><a className="p-span-link" href="https://www.oshawa.ca/en/parks-recreation-and-culture/facilities-and-rentals.aspx" target="_blank" rel="noopener noreferrer">Facility Bookings.</a></span></p>
                 </section>
                 <section className="page-subsection u-content-width">
-                    <h2 className="page-subsection__title--events">Events Happening in Oshawa</h2>
+                    <h2 className="page-subsection__title mb-8">Events Happening in Oshawa</h2>
                     <p>To check out events happening at Ed Broadbent Waterfront Park or other City events, visit the City <span><a className="p-span-link" href="https://calendar.oshawa.ca/" target="_blank" rel="noopener noreferrer">Events Calendar.</a></span></p>
                 </section>
