+++ src/app/(pages)/orangeGarden/page.js
@@ -1,8 +1,8 @@
 "use client";
 
 import "/src/app/globals.css";
-import "/src/app/css/internal.css";
-import "/src/app/css/orangeGarden.css";
+import "/src/app/css/pages/internal.css";
+import "/src/app/css/pages/orangeGarden.css";
 
 import Link from "next/link";
 import Image from "next/image";
