--- src/app/css/events.css
@@ -1,5 +1,0 @@
-.page-subsection__title--events {
-    font-size: var(--font-size--medium);
-    font-weight: 700;
-    margin-bottom: 8px;
-}
