--- src/app/css/gardenOfHumanRights.css
@@ -1,116 +1,0 @@
-.human-rights__list{
-    columns: 2;
-    list-style-position: inside;
-    list-style: none;
-    counter-reset: item;
-    margin-top: 28px;
-    margin-bottom: 32px;
-}
-
-.human-rights__item{
-    counter-increment: item;
-    margin-bottom: 12px;
-}
-
-.human-rights__item:last-child{
-    margin-bottom: 0;
-}
-
-.human-rights__item::before{
-    border: 1px solid var(--color--primary);
-    border-radius: 5px;
-    content: counter(item);
-    color: var(--color--primary);
-    width: 22px;
-    height: 22px;
-    text-align: center;
-    display: inline-block;
-
-    margin-right: 16px;
-
-}
-
-.hr-learn-more{
-    display: flex;
-    margin-top: 20px;
-}
-
-.hr-learn-more__icon{
-    width: 22px;
-    height: 22px;;
-}
-
-.hr-learn-more__description{
-    margin-left: 16px;
-}
-
-/* Modal Background Overlay */
-.modal-overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 1000;
-    backdrop-filter: blur(2px);
-}
-
-/* Fade-in animation */
-@keyframes fadeIn {
-    from {
-        opacity: 0;
-        transform: scale(0.9);
-    }
-    to {
-        opacity: 1;
-        transform: scale(1);
-    }
-}
-
-/* Modal Container */
-.modal-container-general {
-    position: relative;
-    background: white;
-    border-radius: 12px;
-    max-width: 80vw;
-    max-height: 80vh;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    overflow: hidden;
-    box-shadow: 0px 6px 15px rgba(0, 0, 0, 0.5);
-
-    opacity: 0;
-    transform: scale(0.95);
-    animation: fadeIn 0.3s ease-in-out forwards;
-}
-
-/* Modal Image */
-.modal-image {
-    width: auto;
-    height: auto;
-    max-width: 90vw;
-    max-height: 80vh;
-    border-radius: 10px;
-    display: block;
-}
-
-/* Close Button */
-.modal-close-button {
-    position: absolute;
-    top: 10px;
-    right: 10px;
-    background: none;
-    border: none;
-    cursor: pointer;
-    width: 30px;
-    height: 30px;
-}
-
-.modal-close-button img {
-    width: 30px;
-    height: 30px;
-}
