--- src/app/css/general.css
@@ -1,28 +1,0 @@
-.back-to-home-general{
-    margin-top: 72px;
-}
-
-/** ICONS **/
-.amenity-accessibility{
-    background-image: url("/Ed-Broadbent-Waterfront-Park/images/svgs/icons/accessibility.svg");
-}
-
-.amenity-bike {
-    background-image: url('/Ed-Broadbent-Waterfront-Park/images/svgs/icons/bike.svg');
-}
-
-.amenity-garden {
-    background-image: url('/Ed-Broadbent-Waterfront-Park/images/svgs/icons/garden.svg');
-}
-
-.amenity-parking {
-    background-image: url('/Ed-Broadbent-Waterfront-Park/images/svgs/icons/parking.svg');
-}
-
-.amenity-trail {
-    background-image: url('/Ed-Broadbent-Waterfront-Park/images/svgs/icons/recreational-trail.svg');
-}
-
-.back-to-home-general {
-    margin-top: 72px;
-}
