--- src/app/css/internal.css
@@ -1,238 +1,0 @@
-.p-span{
-    color: var(--color--primary);
-    font-weight: 700;
-}
-
-.p-span-bold{
-    font-weight: 700;
-}
-
-.p-span-link{
-    color: var(--color--primary);
-    font-weight: 700;
-    text-decoration: underline;
-}
-
-
-
-/**** For the internal page hero banner ****/
-.page-banner {
-    width: 100%;
-    height: 317px;
-    background-color: rgb(0, 102, 255);
-}
-
-.page-banner__image{
-    object-fit: cover;
-    width: 100%;
-    height: 100%;
-}
-
-
-
-
-/***  Page Section ***/
-
-.page-section{
-    position: relative;
-    background-color: rgb(255, 255, 255);
-    height: auto;
-    border-radius: 27px;
-    margin-top: -33px;
-    padding-top: 64px;
-}
-
-.page-section__title{
-    position: absolute;
-    left: 0;
-    right: 0;
-    margin: auto;
-    top: -22.5px;
-    /* z-index: 9; */
-
-    background-color: var(--color--secondary);
-    color: white;
-    font-size: var(--font-size--internal-title);
-    text-align: center;
-
-    width: 280px;
-    height: 55px;
-
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    border-radius: 50px;
-}
-
-/* .back-home-btn {
-    width: fit-content;
-    display: flex;
-    cursor: pointer;
-    background-color: var(--color--primary);
-    border: none;
-    font-size: 1.5rem;
-    font-weight: 700;
-    padding: 0.6rem 2rem;
-    margin: 73px auto 141px auto;
-    text-align: center;
-    color: white;
-    border-radius: 10px;
-    transition: background-color 0.3s ;
-}
-
-.back-home-btn:active {
-    background-color: #D38621;
-} */
-
-.did-you-know__list {
-    font-size: 1rem;
-    padding-left: 1.5rem;
-    list-style-type: none;
-}
-
-.did-you-know__list li {
-    margin-bottom: 28px; 
-    position: relative;
-}
-
-.did-you-know__list li::before {
-    content: "•";
-    font-size: 2rem;
-    color: var(--color--primary);
-    font-weight: bold;
-    position: absolute;
-    left: -1.6rem;
-    top: -0.7rem;
-}
-
-/**** PAGE SUBSECTIONS ****/
-
-.page-subsection{
-    margin-top: 27px;
-    margin-bottom: 27px;
-}
-
-.page-subsection__title{
-    font-size: var(--font-size--xsmall);
-    margin-bottom: 20px;
-}
-
-.page-subsection__list {
-    font-size: var(--font-size--xsmall);
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    grid-template-rows: repeat(3, auto);
-    column-gap: 3rem;
-    row-gap: 1rem;
-    padding-left: 0;
-    list-style-type: none;
-    padding-bottom: 11px;
-}
-
-.page-subsection__list li {
-    display: flex;
-    align-items: center;
-    gap: 0.8rem;
-}
-
-/* Icon Base Style */
-.amenity-icon {
-    width: 1.6rem;
-    height: 1.6rem;
-    background-size: contain;
-    background-repeat: no-repeat;
-    background-position: center;
-    display: inline-block;
-    border-radius: 0.5rem;
-    padding: 0.3rem;
-}
-
-.back-to-home{
-    display: flex;
-    background-color: var(--color--primary);
-    width: fit-content;
-    font-size: var(--font-size--medium);
-    font-weight: 700;
-    margin: 36px auto 140px auto;
-    padding: 7px 29px;
-    border-radius: 10px;
-    text-align: center;
-}
-
-
-/* Modal Background Overlay */
-.modal-overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 1000;
-    backdrop-filter: blur(2px);
-}
-
-/* Fade-in animation */
-@keyframes fadeIn {
-    from {
-        opacity: 0;
-        transform: scale(0.9);
-    }
-    to {
-        opacity: 1;
-        transform: scale(1);
-    }
-}
-
-/* Modal Container */
-.modal-container-general {
-    position: relative;
-    background: white;
-    border-radius: 12px;
-    max-width: 80vw;
-    max-height: 80vh;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    overflow: hidden;
-    box-shadow: 0px 6px 15px rgba(0, 0, 0, 0.5);
-
-    opacity: 0;
-    transform: scale(0.95);
-    animation: fadeIn 0.3s ease-in-out forwards;
-}
-
-/* Modal Image */
-.modal-image {
-    width: auto;
-    height: auto;
-    max-width: 90vw;
-    max-height: 80vh;
-    border-radius: 10px;
-    display: block;
-}
-
-/* Close Button */
-.modal-close-button {
-    position: absolute;
-    top: 10px;
-    right: 10px;
-    background: none;
-    border: none;
-    cursor: pointer;
-    width: 30px;
-    height: 30px;
-}
-
-.modal-close-button img {
-    width: 30px;
-    height: 30px;
-}
-
-
-
-
