+++ src/app/css/pages/aboutEdBroadbent.css
@@ -3,6 +3,10 @@
     height: auto;
 }
 
+.about-subsection__title{
+    font-size: var(--font-size--medium);
+}
+
 .about-subsection__p{
     border-left: 3px solid var(--color--primary);
     padding-top: 10px;
