+++ src/app/globals.css
@@ -63,6 +63,15 @@ p {
     margin-right: 7%;
 }
 
+.mb-8{
+    margin-bottom: 8px;
+}
+
+.mb-16{
+    margin-bottom: 16px;
+}
+
+
 /**** HEADER *******/
 .nav {
     position: absolute;
